requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
firebase-admin
Pillow